changed lat and lng to decimal instead of float

+++ backend/db/migrations/20230208004318-create-spot.js
@@ -40,10 +40,10 @@ module.exports = {
           allowNull: false,
         },
         lat: {
-          type: Sequelize.FLOAT,
+          type: Sequelize.DECIMAL,
         },
         lng: {
-          type: Sequelize.FLOAT,
+          type: Sequelize.DECIMAL,
         },
         name: {
           type: Sequelize.STRING,
